[Social Aspect][Notifications] Moved the social models/managers into separate folders and files for greater readability
